* OK, so my "refactoring" caused all manner of build errors in the DBS --> Let this be a lesson to you all. Not building/fixing before you commit your code is not big and its not clever.

+++ alitheia/db/src/eu/sqooss/service/db/FileGroup.java
@@ -47,11 +47,11 @@ public class FileGroup {
 	// Nothing to do
     }
 
-    public int getId() {
+    public long getId() {
 	return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
 	this.id = id;
     }
 
@@ -95,11 +95,11 @@ public class FileGroup {
 	this.lastUsed = lastUsed;
     }
 
-    public Long getProjectVersion() {
+    public long getProjectVersion() {
 	return projectVersion;
     }
 
-    public void setProjectVersion(Long projectVersion ) {
+    public void setProjectVersion(long projectVersion ) {
 	this.projectVersion = projectVersion;
     }
 }
