+++ alitheia/db/src/eu/sqooss/service/db/Measurement.java
@@ -45,27 +45,27 @@ public class Measurement {
 	// Nothing to do here
     }
 
-    public int getId() {
+    public long getId() {
 	return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
 	this.id = id;
     }
 
-    public int getMetric() {
+    public long getMetric() {
 	return metric;
     }
 
-    public void setMetric(int metric) {
+    public void setMetric(long metric) {
 	this.metric = metric;
     }
 
-    public Long getProjectVersion() {
+    public long getProjectVersion() {
 	return projectVersion;
     }
 
-    public void setProjectVersion(Long projectVersion) {
+    public void setProjectVersion(long projectVersion) {
 	this.projectVersion = projectVersion;
     }
