+++ alitheia/db/src/eu/sqooss/service/db/MetricType.java
@@ -40,11 +40,11 @@ public class MetricType {
 	// Nothing to do here
     }
 
-    public int getId() {
+    public long getId() {
 	return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
 	this.id = id;
     }
