+++ alitheia/db/src/eu/sqooss/service/db/Plugin.java
@@ -41,19 +41,19 @@ public class Plugin {
 	// Nothing to do here
     }
 
-    public int getId() {
+    public long getId() {
 	return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
 	this.id = id;
     }
 
-    public int getMetric() {
+    public long getMetric() {
 	return metric;
     }
 
-    public void setMetric(int metric) {
+    public void setMetric(long metric) {
 	this.metric = metric;
     }
